+++ scripts/_migration.mjs
@@ -8,28 +8,24 @@ Hooks.once('ready', async () => {
 		[`${MODULE.ID}`]: {
 			name: MODULE.TITLE,
 			theme: {
-				// v0.3.2 Them Options
+				// Generic Windows
 				libraryFantasyRPGUIGenericWindow: {value: 'true'},
-				libraryFantasyRPGUIGenericWindowInputs: {value: 'true'},
 				libraryFantasyRPGUIGenericWindowHideHeaderButtons: {value: 'true'},
+				libraryFantasyRPGUIGenericWindowInputs: {value: 'true'},
+				// Interface
 				libraryFantasyRPGUIInterface: {value: 'true'},
 				libraryFantasyRPGUIInterfaceControls: {value: 'true'},
 				libraryFantasyRPGUIInterfaceHotbar: {value: 'true'},
 				libraryFantasyRPGUIInterfaceNavigation: {value: 'true'},
 				libraryFantasyRPGUIInterfacePlayers: {value: 'true'},
-				libraryFantasyRPGUIDialogs: {value: 'true'},
-				libraryFantasyRPGUIJournals: {value: 'true'},
-				libraryFantasyRPGUIJournalsHideHeaderButtons: {value: 'true'},
+				// Sidebar
 				libraryFantasyRPGUISidebar: {value: 'true'},
 				libraryFantasyRPGUISidebarNav: {value: 'true'},
 				libraryFantasyRPGUISidebarChat: {value: 'true'},
 				libraryFantasyRPGUISidebarChatMessages: {value: 'true'},
+				libraryFantasyRPGUISidebarCombat: {value: 'true'},
 				libraryFantasyRPGUISidebarScenes: {value: 'true'},
 				libraryFantasyRPGUISidebarScenesCompact: {value: 'true'},
-				libraryFantasyRPGUISidebarSettings: {value: 'true'},
-				libraryFantasyRPGUIModulesFoundryTaskbar: {value: 'true'},
-				// v0.3.3 Them Options
-				libraryFantasyRPGUISidebarCombat: {value: 'true'},
 				libraryFantasyRPGUISidebarActors: {value: 'true'},
 				libraryFantasyRPGUISidebarItems: {value: 'true'},
 				libraryFantasyRPGUISidebarJournal: {value: 'true'},
@@ -37,63 +33,88 @@ Hooks.once('ready', async () => {
 				libraryFantasyRPGUISidebarCards: {value: 'true'},
 				libraryFantasyRPGUISidebarPlaylists: {value: 'true'},
 				libraryFantasyRPGUISidebarCompendium: {value: 'true'},
-				libraryFantasyRPGUIProseMirror: {value: 'true'}
+				libraryFantasyRPGUISidebarSettings: {value: 'true'},
+				// Themed Windows
+				libraryFantasyRPGUIDialogs: {value: 'true'},
+				libraryFantasyRPGUIJournals: {value: 'true'},
+				libraryFantasyRPGUIJournalsHideHeaderButtons: {value: 'true'},
+				libraryFantasyRPGUIProseMirror: {value: 'true'},
+				// Systems
+				libraryFantasyRPGUISystemPF2e: {value: 'true'},
+				// Modules
+				libraryFantasyRPGUIModulesAlwaysHP: {value: 'true'},
+				libraryFantasyRPGUIModulesFoundryTaskbar: {value: 'true'},
+				libraryFantasyRPGUIModulesWindowControls: {value: 'true'}
 			}
 		}
 	};
 
+	// If User is Player and Master Theme is Enabled, Return
+	if (!game.user.isGM && game.settings.get('lib-themer', 'enableMasterTheme')) return false;
+
 	// Check if Preset Exists, if Not Ask to Create and Enable It
 	if (!presets.hasOwnProperty(MODULE.ID)) {
-		Dialog.confirm({
-			id: `${MODULE.ID}-create-preset`,
-			title: MODULE.TITLE,
-			content: `<p style="margin-top: 0px;">${MODULE.localize('dialog.createPreset')}</p>`,
-			yes: (elemDialog) => {
-				game.settings.set('lib-themer', 'presets', foundry.utils.mergeObject(game.settings.get('lib-themer', 'presets'), rpgUITheme, { inplace: false })).then(response => {
-					game.settings.set('lib-themer', 'themeSettings', game.settings.get('lib-themer', 'presets')[MODULE.ID]).then(response => {
-						for (const [property, value] of Object.entries(game.settings.get('lib-themer','themeSettings').theme)) {
+		async function createPreset() {
+			return Dialog.confirm({
+				id: `${MODULE.ID}-create-preset`,
+				title: MODULE.TITLE,
+				content: `<p style="margin-top: 0px;">${MODULE.localize('dialog.createPreset')}</p>`,
+				yes: async (elemDialog) => {
+					if (game.user.isGM) {
+						await game.settings.set('lib-themer', 'presets', foundry.utils.mergeObject(game.settings.get('lib-themer', 'presets'), rpgUITheme, { inplace: false }));
+					}
+
+					// Update Current Settings
+					let currentThemeSettings = game.settings.get('lib-themer', 'themeSettings');
+					currentThemeSettings = foundry.utils.mergeObject(currentThemeSettings, rpgUITheme[MODULE.ID].theme, { inplace: false });
+
+					// Apply Current Settings
+					game.settings.set('lib-themer', 'themeSettings', currentThemeSettings).then(response => {
+						for (const [property, value] of Object.entries(currentThemeSettings)) {
 							game.modules.get('lib-themer').API.setCSSVariable(property, value.value);
 						}
 					});
-				})
-			}, no: () => {
-				game.settings.set('lib-themer', 'presets', foundry.utils.mergeObject(game.settings.get('lib-themer', 'presets'), rpgUITheme, { inplace: false }))
-			}
-		})
+				}, no: () => {
+					if (game.user.isGM) {
+						game.settings.set('lib-themer', 'presets', foundry.utils.mergeObject(game.settings.get('lib-themer', 'presets'), rpgUITheme, { inplace: false }));
+					}
+				}
+			})
+		}
+		createPreset();
 	}else{
-		// Preset Exists Check if Its updated
-		let preset = presets[MODULE.ID].theme;
+		if ((Object.keys(rpgUITheme[MODULE.ID].theme)?.length ?? 0) > (Object.keys(presets[MODULE.ID].theme)?.length ?? 0)) {
+			async function updatePreset() {
+				// Get Current Preset and Merge Update Into It.
+				rpgUITheme[MODULE.ID].theme = foundry.utils.mergeObject(rpgUITheme[MODULE.ID].theme, presets[MODULE.ID].theme, { inplace: false, insertKeys: false });
 
-		// Version 0.3.3
-		if (!presets[MODULE.ID].theme.hasOwnProperty('libraryFantasyRPGUISidebarActors')) {
-			let v033Options = {
-				libraryFantasyRPGUISidebarCombat: {value: 'true'},
-				libraryFantasyRPGUISidebarActors: {value: 'true'},
-				libraryFantasyRPGUISidebarItems: {value: 'true'},
-				libraryFantasyRPGUISidebarJournal: {value: 'true'},
-				libraryFantasyRPGUISidebarTables: {value: 'true'},
-				libraryFantasyRPGUISidebarCards: {value: 'true'},
-				libraryFantasyRPGUISidebarPlaylists: {value: 'true'},
-				libraryFantasyRPGUISidebarCompendium: {value: 'true'},
-				libraryFantasyRPGUIProseMirror: {value: 'true'}
-			}
+				return Dialog.confirm({
+					id: `${MODULE.ID}-update-preset`,
+					title: MODULE.TITLE,
+					content: `<p style="margin-top: 0px;">${MODULE.localize('migration.updatePreset')}</p>`,
+					yes: async (elemDialog) => {
+						if (game.user.isGM) {
+							await game.settings.set('lib-themer', 'presets', foundry.utils.mergeObject(game.settings.get('lib-themer', 'presets'), rpgUITheme, { inplace: false }));
+						}
 
-			Dialog.confirm({
-				id: `${MODULE.ID}-update-preset`,
-				title: MODULE.TITLE,
-				content: `<p style="margin-top: 0px;">${MODULE.localize('migration.v033.updatePreset')}</p>`,
-				yes: (elemDialog) => {
-					game.settings.set('lib-themer', 'presets', foundry.utils.mergeObject(rpgUITheme[MODULE.ID], presets[MODULE.ID], { inplace: false })).then(response => {
-						game.settings.set('lib-themer', 'themeSettings', foundry.utils.mergeObject(v033Options, game.settings.get('lib-themer', 'themeSettings'), { inplace: false })).then(response => {
-							for (const [property, value] of Object.entries(v033Options)) {
+						// Update Current Settings
+						let currentThemeSettings = game.settings.get('lib-themer', 'themeSettings');
+						currentThemeSettings = foundry.utils.mergeObject(currentThemeSettings, rpgUITheme[MODULE.ID].theme, { inplace: false });
+	
+						// Apply Current Settings
+						game.settings.set('lib-themer', 'themeSettings', currentThemeSettings).then(response => {
+							for (const [property, value] of Object.entries(currentThemeSettings)) {
 								game.modules.get('lib-themer').API.setCSSVariable(property, value.value);
 							}
 						});
-					})
-				}, no: () => {
-					game.settings.set('lib-themer', 'presets', foundry.utils.mergeObject(rpgUITheme[MODULE.ID], presets[MODULE.ID], { inplace: false }));
-				}
-			});
+					}, no: () => {
+						if (game.user.isGM) {
+							game.settings.set('lib-themer', 'presets', foundry.utils.mergeObject(game.settings.get('lib-themer', 'presets'), rpgUITheme, { inplace: false }));
+						}
+					}
+				})
+			}
+			updatePreset();
 		}
 	}
 });