+++ CHANGELOG.md
@@ -1,6 +1,15 @@
+# Version 0.3.6 - Compatibility Patches
+- Added Basic Support to Pathfinder 2e System
+  - Character Sheets
+  - Heading for Settings Sidebar
+- Added Support for Window Controls
+- Fixed Compendium and Settings Sidebar Headings from staying fixed while scrolling through content
+- Removed Styling from Monster Blocks
+- Changed Migration Logic to handle updating Preset only if GM
+
 # Version 0.3.5 - Compatibility Patches
-- Removed Styling from About Time (Will try to theme it later)
-- Removed Styling from Simple Calendar (Will try to theme it later)
+- Removed Styling from About Time
+- Removed Styling from Simple Calendar
 - Added Support for Always HP
 - Added Support for Pop Out
   - Header Buttons will not be seen in the PopOut. Click on the couplings will trigger the button, but there is no way to see the text of the button at the moment.